fix parser generator
